GUI optimization + more formatting
Also renamed FPSTimer to FPSTracker as I felt it was more clear.

RAMTracker GetMegabytesFromBytes function now returns a float instead of an int

+++ OpenGLBasics/RAMTracker.h
@@ -15,8 +15,8 @@ public:
 
     static void AddCurrentRAMValue();
 
-    static size_t GetMegabytesFromBytes(size_t bytes) {
-        return (bytes / 1000 / 1000);
+    static float GetMegabytesFromBytes(size_t bytes) {
+        return bytes / 1000.0f / 1000.0f;
     }
 };
